Avoid sending expected CleverTap errors to sentry (#2573)
### Description

Sentry is getting a lot of errors for `CleverTap initialUrl is nil` and `CleverTap initialUrl is null` this is expected when the app launches normally and not from a deep link. These are now warnings and not sent to sentry. Once live these issues should not be seen in sentry [CleverTap InitialUrl is null](https://sentry.io/organizations/valora-inc/issues/3020003220/?query=is%3Aignored&sort=user&statsPeriod=14d) and [CleverTap initialUrl is nil](https://sentry.io/organizations/valora-inc/issues/3020505928/?query=is%3Aignored&sort=user&statsPeriod=14d).

### Other changes

- Use regex to test error and adjusted matching.

### Tested

Tested locally on iOS. 

### How others should test

This does not need to be tested by QA.

### Related issues

- Fixes #2350

### Backwards compatibility

Yes