chore: remove geth logs from Logger and combine with ReactNativeLogger (#2615)
### Description

Title says it all.

There was no need to have a separate `Logger` and `ReactNativeLogger` anymore. Only `Logger` remains.

Also this should help address the issue when exporting big log files. The data doesn't need to go through the RN bridge to create a combined log file anymore. We just use the log path. 
On Android, we're copying the file to a world readable dir, but I believe this extra step can be removed too.

### Tested

- Updated tests
- Manually on iOS and Android

### How others should test

No need to test this.

### Related issues

- Related to RET-256 and RET-336

### Backwards compatibility

Yes