Add placeholder image if no dapp logo can be displayed (#2635)
### Description

For dapps using dappkit and not listed on the dapps list, there is no image that we can display on the connection/transaction request bottom sheets. This PR adds a default placeholder image (inspired by the send screen)

I tried to use the favicons but found with the first 2 dapps i tried (revo, moola) that their favicons were already located in different urls (not the standard `/favicon.ico`) so i gave up. a quick google search said that around 40% of websites don't have a favicon so wasn't sure that the favicon fallback would be fullproof anyway. 

![Simulator Screen Shot - iPhone 11 - 2022-06-23 at 11 38 05](https://user-images.githubusercontent.com/20150449/175268922-613370d9-d91f-4fa5-9ba8-190e0b2c5b71.png)


### Other changes

N/A

### Tested

N/A

### How others should test

N/A

### Related issues

- Relates #2342

### Backwards compatibility

N/A