Add react-navigation-bottom-sheet for dapp connection (RET-240) (#2640)
### Description

As the title, this PR adds a new lib to make the bottom sheets native. This change is only applied to the WC/dappkit bottom sheets for now.

To note:
- the cross button to dismiss bottom sheets was removed because now we can swipe to dismiss
- i added a (potentially brittle?) function to detect if the bottom sheet is there, because we need to know if a user has dismissed the bottom sheet by selecting the "cancel" button OR by swiping to dismiss the sheet. In both situations, we want to call the "onDeny" callback but for the latter scenario we do not need to navigate back to dismiss the sheet anymore.


https://user-images.githubusercontent.com/20150449/175883092-db3bbb21-e2fa-44e4-8a3d-4165275dc29d.mp4


### Other changes

N/A

### Tested

Manually

### How others should test

Ensure that you can connect to dapps and transact with dapps. Please try with dapps using WalletConnect (e.g. ChurritoFi, Mento) and dappKit (e.g. Revo, Moola)

### Related issues

- Fixes RET-240
- Fixes #2342 

### Backwards compatibility

N/A