chore(drawer-menu): use swap screen remote config for celo menu item position (#2716)
### Description

From Ellen's feedback from demo, seems like we also want to control the CELO drawer menu item position with the same remote config flag as the swap menu item. This was missed in the requirements.

### Other changes

N/A

### Tested

Manually

### How others should test

The CELO item should be the second drawer menu item (i.e. on Mainnet). If the swap menu item is displayed (i.e. in Alfajores), then the CELO item should be after "add & withdraw". 

### Related issues

- Relates to RET-319

### Backwards compatibility

Yes