fix(homefeed): ensure that users can scroll to load more (#2779)
### Description

Context: https://valora-app.slack.com/archives/C018VLA3YAK/p1660091022594799

The solution chosen for this issue is to continue to fetch transactions from the client until a minimum threshold is met (currently 10) so that the section list is filled and the user can scroll to load more.

In this PR:
- To ensure no weird bugs and avoid concurrency issues, i refactored the state variables to keep the cumulative transactions + pageInfo in the same state variable so that they are always in sync. 
- Added a hook to refetch transactions until we have 10, and refetch transactions if the returned page is empty
- Updated tests

### Other changes

N/A

### Tested

Manually and with unit tests

### How others should test

Please try this with many accounts that have varying number of transactions - load the app (from a fresh install, and from cold start) and ensure that the home feed displays transactions. If the account has many transactions, ensure that you can scroll down to load more transactions.

### Related issues

N/A

### Backwards compatibility

Yes