fix: navigating to connect phone number with existing wallet (#2777)
### Description

When navigating to the connect phone number with an existing account users were being served 'Failed to create your wallet' screen due to a dispatch to initialize wallet. This PR checks via the walletAddressSelector returns null prior to attempting to initialize the wallet.

### Other changes

N/A

### Tested

Tested locally on iOS.

### How others should test

1. Re-install with new account.
2. Restart-app.
3. Attempt to connect phone number from the home card, settings, or supercharger card.
4. Pin Entry is not served.
5. Tapping back takes user back to the previous screen and failed to create your wallet screen is not served.

### Related issues

- Fixes #2511

### Backwards compatibility

Yes