feat: feature flag for guided onboarding copy (#2817)
### Description

An additional remote config feature flag, `showGuidedOnboardingCopy` , is added and recognized through changes made in this PR. This boolean defaults to `false`; application would appear identical to that before this PR when this boolean is `false`.
Different header titles and guide texts are shown if the new feature flag is set to `true` from Firebase. For the Welcome page, the welcome prompt is changed. 
The pages affected are the Welcome, Name and Picture, PIN Setup and Biometrics Setup.

### Other changes

None

### Tested

Tests for individual pages that were affected by this flag were written, with the selector mocked to return true on that flag, and affected copies verified to be the "guided onboarding" ones.
This feature flag currently has only been added to alfajores, but not mainnet yet.

### How others should test

Depending on how the feature flag is applied, test users should see different copy during onboarding.

### Related issues

- Fixes [ACT-417](https://linear.app/valora/issue/ACT-417/onboarding-experiment-trusted-guide)

### Backwards compatibility

Backwards compatible in that an older app that do not recognize the new feature flag would ignore it. Upgrade compatible because default value of false will be added during upgrading the old state.

Co-authored-by: Charlie Andrews-Jubelt <cajubelt@users.noreply.github.com>