feat: feature flag for guided onboarding copy (#2850)
### Description

This PR is a re-visit to #2817 which was reverted in #2840 due to a bug
in the Pincode behavior, causing it to change even when the feature flag
was set to false; this PR includes the same features with this bug
addressed.

In addition, during onboarding, copy and design had an update for the
"re-enter pin" page, as shown in the screenshot and in the [design
file](https://www.figma.com/file/j84MQn5xaOJ9RBjuIZgxPN/Onboarding-Experiments?node-id=1852%3A16671).

Original description: an additional remote config feature flag,
`showGuidedOnboardingCopy` , is added and recognized through changes
made in this PR. This boolean defaults to `false`; application would
appear identical to that before this PR when this boolean is `false`.
Different header titles and guide texts are shown if the new feature
flag is set to `true` from Firebase. For the Welcome page, the welcome
prompt is changed.
The pages affected are the Welcome, Name and Picture, PIN Setup and
Biometrics Setup.

### Other changes

None

### Tested

Tests for individual pages that were affected by this flag were written,
with the selector mocked to return true on that flag, and affected
copies verified to be the "guided onboarding" ones.
This feature flag currently has only been added to alfajores, but not
mainnet yet.

### How others should test

For engineers trying it out on local build, you can edit the
`src/app/selectors.ts:132` to return `true` without setting feature flag
in Firebase.
Depending on how the feature flag is applied, test users should see
different copy during onboarding.

### Related issues

- Fixes
[ACT-417](https://linear.app/valora/issue/ACT-417/onboarding-experiment-trusted-guide)

### Backwards compatibility

Backwards compatible in that an older app that do not recognize the new
feature flag would ignore it. Upgrade compatible because default value
of false will be added during upgrading the old state.