feat(fiatconnect): Check KYC after submitting account (#2912)
### Description
Fixes ACT-384. This PR adds checks for KYC status after a user
successfully submits a fiat account on the FiatDetails page, when KYC is
required for the quote the user has selected. Depending on the KYC
status, we either navigate to the Review screen, or a particular KYC
status screen. In the exact case where KYC is required _and_ it has been
approved by the time the user has submitted their account, we show a
green checkmark to the user to indicate their ID verification success.

### Other changes

_Describe any minor or "drive-by" changes here._

### Tested

Unit tested. Not easily end-to-end testable at the moment.

### How others should test

_Does this need to be tested by QA in the next release cycle? If so
please give a brief explanation of how to test these changes._

### Related issues

Fixes ACT-384.

### Backwards compatibility

_Brief explanation of why these changes are/are not backwards
compatible._