feat: Add share URL flow (#2886)
### Description

This PR adds a new drawer navigation item "Invite". Users can open the
Invite screen to get a deep link directing other users to the app.

### Other changes

- Grabbing the share link kicks off an analytics event and on iOS, the
results of the share (if it was abandoned, or what app was used for the
sure) are attached as properties. On Android the results are not
available.
- Users opening the app via the invite link also kicks off an analytics
event, with the inviter's wallet address as a property.

### Tested

These steps were followed on each platform:

Link creation:
1. Click the `Invite` link in the drawer navigator
2. The `Invite` button should initially be disabled, as the link is
being created
3. When the `Invite` button is enabled, click it.
4. The native share drawer should appear, with a short message including
the share link.

Clicking the link:
- If on a mobile device and the app is not installed, should open
valoraapp.com
 - If on a desktop, should open valoraapp.com
 - If on a mobile device and the app is installed, should open the app

### How others should test

I imagine we'll want to add this to QA test steps.

### Related issues

- Fixes EA-300

### Backwards compatibility

This PR should be backwards compatible. Link creation hinges off a
remote config variable.


### Relevant UI:


https://user-images.githubusercontent.com/8698364/192322143-74af2dd9-80b8-44d4-98df-0f5c3077f697.mp4



https://user-images.githubusercontent.com/8698364/192322151-bb9bac8a-d2c6-49a7-8388-c7b83c15fadf.mp4