fix: incorrect connection error on WalletConnect connection (#2940)
### Description

When a user manually scans a QR code, during the time between when the
QR code is first detected and when the connection bottom sheet appears
and is acknowledged, other requests to handle the same code are
dispatched. This causes the behaviour where after the connection is
successful, another connection bottom sheet appears and seems to load
and time out.

This PR tries to prevent such duplicate requests from being fired. 

### Other changes

N/A

### Tested

Manually

### How others should test

On a physical device, attempt to connect to a dapp on a browser by
scanning the QR code. The connection should be successful.

### Related issues

- Fixes RET-372

### Backwards compatibility

Yes