fix(fiatconnect): Fix issue when checking KYC after submitting fiat account (#2931)
### Description

TLDR; the switch/case that deals with KYC statuses after posting a fiat
account was falling through when it wasn't supposed to, which resulted
in `navigate` being called several times in a row. The tests didn't
catch it, since they didn't check to make sure that `navigate` was only
being called a _single_ time. Fixed the implementation, as well as
updated the tests to ensure that this sort of thing gets caught.

### Other changes

N/A.

### Tested

Unit tested.

### How others should test

N/A.

### Related issues

N/A.

### Backwards compatibility

N/A.