fix(fiatconnect): Fix KYC info display (#2928)
### Description

Updates the info section for provider quotes to correctly display
whether or not KYC is required. For external providers, this will always
display `ID Required`, but for FiatConnect providers, the string will be
dynamic, based on the _type_ of KYC that's required. (Somewhat
confusingly, the only FC KYC we support is also set to display `ID
Required` at the moment.)

### Tested

Unit/manual testing:

#### KYC Required


![2022-09-29-135036_668x1415_scrot](https://user-images.githubusercontent.com/569401/193106041-914bcb0f-e1db-40e8-9291-f53d780ab786.png)

#### KYC Not Required


![2022-09-29-135057_664x1409_scrot](https://user-images.githubusercontent.com/569401/193106306-1b5020f4-57a8-458c-8e12-918772f77cc8.png)


### How others should test

N/A

### Backwards compatibility

_Brief explanation of why these changes are/are not backwards
compatible._

Co-authored-by: Satish Ravi <satish.ravi@valoraapp.com>
Co-authored-by: Charlie Andrews-Jubelt <cajubelt@users.noreply.github.com>