feat(analytics): use Statsig skip button param in onboarding (#2950)
### Description

This PR makes use of the Statsig experiment params in the
`NameAndPicture` screen. Currently, Statsig is initialized but unused in
the wallet. After merging this PR, the `NameAndPicture` screen will show
a skip button only if the user was bucketed into that trial group by
Statsig.

### Other changes

None

### Tested

Unit tests.

### How others should test

Verify skip button is shown sometimes on the Name Screen when
onboarding. Also, test that airplane mode doesn't lead to issues.

### Related issues

-
[ACT-436](https://linear.app/valora/issue/ACT-436/skip-button-on-name-screen)

### Backwards compatibility

N/A