feat: update wallet to use CPV status (#2946)
### Description

This PR
- add CPV status in redux (use new variable so that the old one can be
used for triggering migration to CPV)
- on CPV success, store phone number and CPV status in redux
- in most places, use verification status from central or decentral
system depending on remote config
- for supercharge and support, use both with an or condition (since
users are eligible for supercharge either way)

### Other changes

N/A

### Tested

In logs sent to support:
- contains correct phone verification status (central and decentral,
separately)

The following is true when CPV is enabled + the user is not verified
centrally or decentrally. The following is also true when CPV is
disabled + the user is not verified decentrally:
- settings screen / supercharge screen / home card / send screen shows
option to connect phone number. on click, it takes you to the CPV screen
(or verification education screen if CPV is off)
- the drawer menu does not show any phone number
- contacts are not imported and cannot be seen on the send screen

The following is true if CPV enabled + user is verified centrally **or**
decentrally. The following is also true when CPV is disabled and the
user is verified decentrally:
- no option to connect phone number on settings screen / supercharge
screen / home card / send screen
- the drawer menu shows verified phone number

**The user will count as "verified" even if CPV is enabled and the user
is verified decentrally only.** After some period of supporting both,
users who have not migrated will need to complete verification again,
but that is in the future.


### How others should test

As above

### Related issues

- Fixes RET-384

### Backwards compatibility

Yes

Co-authored-by: Jean Regisser <jean.regisser@gmail.com>
Co-authored-by: Tom McGuire <Mcgtom10@gmail.com>