feat: add username generator function (#2958)
### Description

For an [onboarding
experiment](https://www.notion.so/valora-inc/Onboarding-Name-Step-4b6ef3485ac64b0caeb87147887f68a5),
we need to be able to generate random, fun usernames. This PR implements
part of the required functionality, adding a list of about 500
adjectives and nouns each that will be used to generate the names.

### Other changes

N/A

### Tested

Added unit test. Will be more throughly tested in a subsequent PR.

### How others should test

N/A

### Related issues

-
[ACT-435](https://linear.app/valora/issue/ACT-435/username-generator-helper-function)

### Backwards compatibility

N/A

Co-authored-by: Charlie Andrews-Jubelt <cajubelt@users.noreply.github.com>
Co-authored-by: Tom McGuire <Mcgtom10@gmail.com>
Co-authored-by: Joe Bergeron <jbergero@alum.mit.edu>