feat: restore phone verification status when restoring wallet (#2980)
### Description

When initialising (restoring) wallet, check if there is a phone number
already verified for this wallet. If so, store the verified phone number
in redux so that the user correctly displays as verified in the app.

Note that the signed message is required for making the address lookup
query, and this is generated potentially after the restore saga finishes
so that's why it's in the initializeAccount saga now...

### Other changes

N/A

### Tested

Manually + unit tests.

### How others should test

Verify a phone number against an address. Reinstall the app and restore
this account - the app should know that your number is verified.

### Related issues

- Fixes RET-451

### Backwards compatibility

Yes

Co-authored-by: Jean Regisser <jean.regisser@gmail.com>