feat: add username blocklist from Statsig dynamic config (#2974)
### Description

For an [onboarding
experiment](https://www.notion.so/valora-inc/Onboarding-Name-Step-4b6ef3485ac64b0caeb87147887f68a5),
we need to be able to generate random, fun usernames. This PR is a
continuation of #2958 which obtains the adjective and noun blocklists
from Statsig dynamic config. It does not yet use the blocklists in the
`NameAndPicture` screen.

### Other changes

N/A

### Tested

N/A

### How others should test

N/A

### Related issues

-
[ACT-435](https://linear.app/valora/issue/ACT-435/username-generator-helper-function)

### Backwards compatibility

N/A