fix: crowdin automerge job sometimes leaves incorrect reviews (#2988)
### Description

I found the automerge job sometimes leaves the "changes requested"
review when the Crowdin PR does not contain any unexpected changes. I
suspect that it is because the review is done before the branch is
properly updated. e.g. https://github.com/valora-inc/wallet/pull/2982

In this PR
- wait for a few seconds after updating the branch, to do the review
- make the logging more explicit so that we can see which files are
causing the incorrect review

### Other changes

N/A

### Tested

Tested
https://github.com/valora-inc/wallet/actions/runs/3243609016/jobs/5318572378

### How others should test

N/A

### Related issues

N/A

### Backwards compatibility

Yes