feat: add background phone verification migration (#3001)
### Description

The migration should be run after remote config fetched if CPV is on +
the user is verified decentrally _but not centrally_. In case of
failures, it should be silent since we will see the errors on the cloud
function logs already and the wallet is backwards compatible.

### Other changes

The app/saga.test.ts file was getting a bit messy with inconsistent
grouping of tests. I rearranged the grouping of the tests, would
recommend reviewing without white spaces. **most of the lines changed in
this PR are from the test file!**

### Tested

As a decentrally verified user, open the app. Verify that you can send
to another number that is only verified centrally. Verify that when
restoring this account, the verification status is preserved
automatically.

### How others should test

As above.

### Related issues

- Fixes RET-405

### Backwards compatibility

Yes

Co-authored-by: Tom McGuire <Mcgtom10@gmail.com>