fix: CPV in restore wallet flow with phone verification (#3016)
### Description

The bug:
The wallet is created / restored at the last step of onboarding. If
phone verification is required at onboarding, when the user arrives at
the phone verification screen in the restore wallet flow, the account
and signed message are not yet available (since they are programmed to
be created after this screen) and therefore attempting to do central
phone verification at this stage will fail.

This PR:
- create the account after the restore screen. at this point, we have
all the information needed to restore the account anyway. we can also
verify the CPV status here, to decide to navigate to the phone
verification screen or not.
- by creating the account after the seed phrase screen, users will have
the wallet address and signed message needed for CPV.
- of course this introduces a delay on the seed phrase screen while we
wait for the account to be created. previously this account creation is
done asynchronously while navigating to the home screen, however the
user will not see anything of value on the home screen while the account
is being created so i think this delay is acceptable.

The one wrinkle:
- At the start of onboarding, we haven't created the account yet
obviously. Therefore, there is no way to know the user's address to
check if they are already verified / should go through phone
verification. So it's not possible to determine the correct number of
onboarding steps to be displayed in the screen header...if a user is
already verified after we restore the wallet using the seed phrase in
step 3/4, they will not need to complete step 4/4 (the phone
verification step). The current flow may lead the user to think the app
glitched and missed a step? Technically I can't think of a way around
this, so may need to follow up with a UI change.


### Other changes

N/A

### Tested

Manually, during the restore wallet flow, you should be able to go
through CPV. When restoring a wallet that is already verified, you
should not see the CPV flow.

### How others should test

As above

### Related issues

N/A

### Backwards compatibility

Yes