feat(fiatconnect): Add ability to navigate to review screen, re-fetching cached quotes (#3036)
### Description

Fixes ACT-474.

This PR adds in the ability to navigate to the `FiatConnectReview`
screen via the new `FiatConnectReviewWrapper` screen, while
automatically fetching a new quote from cached parameters.

There were a few different options for implementing this, but I ended up
going with a small wrapper screen that handles showing a spinner and
dispatching an existing action in order to refetch from a cached quote,
before navigating to the Review screen. Some of the other options...
* Update `FiatConnectReview` screen to only _optionally_ take a quote,
and add additional parameters to tell it to refetch from some particular
cached quote params. Making the `quote` param optional to this component
would have been a mess from a typing perspective, and would further
complicate the already somewhat bloated review screen.
* Create a brand new action/saga to be triggered from the KYC accepted
notification handler directly, which would be responsible for refetching
the quote and navigating to the existing `FiatConnectReview` screen.
This would have required a brand new saga/action, and is further
complicated by the fact that we want to immediately show a spinner page
while the quote is being refetched; not sure how we would do that
without modifying an existing screen or adding a new one.
* There was a potential option of re-using the primary quote selection
saga to handle this, though without larger changes to it, certain
failure/edge cases could have led the user to some relatively unexpected
behavior, considering that the logic in this PR is meant to be triggered
after pressing a push notification/homecard.

The chosen solution reuses the quote refetching logic, which needed a
slight refactor. Currently, when we refetch quotes, we already know what
fiat account the user has selected. When fetching from a cached quote
however, we will not have this luxury, and so must check to see if any
of the quote(s) returned from the provider during the re-fetch match
with an account the user has on file. As a result, `fiatAccount` is now
an _optional_ parameter to `RefetchQuoteAction`. If it is not provided,
we check cached fiat account uses after fetching the quote, check for a
match between the quote's constraints and our cached accounts, and if a
match exists, fetch the actual account information from the provider.

### Other changes

One inconsistency I noticed while making these changes was that when
passing fiat account information through different screens in route
params, they're typed as `ObfuscatedFiatAccountData`, while in
actuality, they are _always_ of type `FiatAccount`. This worked, since
`FiatAccount` is just an intersection of `ObfuscatedFiatAccountData` and
a `providerId` field. I've updated our usage of fiat accounts in these
places to be typed as `FiatAccount`, as it (probably?) should have been
already.

### Tested

Unit tested.

### How others should test

Not integrated/triggered in-app currently.

### Related issues

- Fixes ACT-474

### Backwards compatibility

N/A.

Co-authored-by: Charlie Andrews-Jubelt <cajubelt@users.noreply.github.com>