feat(fiatconnect): handle transfer in requests (#3114)
### Description

Refactored `handleCreateFiatConnectTransfer` to create two helper
functions
* `_initiateTransfer` for starting a fiatconnect transfer regardless of
in/out
* `_initiateTransaction` for initiating the transaction on blockchain
which only happens for transfers out

Then added some small functionality to only call `_initiateTransaction`
for transfers out.

#### Video

Ignore the copy on the Review screen. It is not ready yet and in a
separate PR


https://user-images.githubusercontent.com/8432644/202314313-2a7c2171-6298-4eff-b640-219aefc28bcb.mp4


### Tested

Added quite a few more tests and improved the existing ones. Also tested
out the flow on device


### How others should test

N/a

### Related issues

https://linear.app/valora/issue/ACT-406/initiate-transfer-in

Co-authored-by: jh2oman <jacobwaterman@MacBook-Pro.local>