feat(fiatconnect): add mobile money to fiat details screen (#3117)
### Description

Currently, the Fiat Details screen does not have any support for a
Mobile Money account. This PR allows Mobile Money accounts to be added
on the Fiat Details screen. There isn't much phone number validation
yet, just a check that the number is at most 15 digits as per the
[international standards](https://en.wikipedia.org/wiki/E.164).


https://user-images.githubusercontent.com/20508929/202808974-e5ebbd04-e7c0-4849-b9e5-5cb8417a0a01.mp4


### Other changes

N/A

### Tested

Added unit tests.

### How others should test

N/A, Mobile Money is not yet supported in the wallet.

### Related issues

- Fixes ACT-516

### Backwards compatibility

N/A