fix: add swap fee estimate approximation (#3120)
### Description

This PR adds a fee estimate to swaps so that swap amounts using the
"max" button do not receive the "transfer exceeded balance" error. It's
difficult to do this correctly when we don't have the swap transaction,
have opted to go with a multiple of a simple transfer fee for now. This
means that users using the "max" button will be much less likely to see
their swap rejected.

### Other changes

Since we are not using escrow invites anymore, rather than make the fee
calculation function more complicated I opted to break the invite fee
estimation....

### Tested

Made a successful swap using the max button

### How others should test

As above

### Related issues

- Fixes RET-495

### Backwards compatibility

Yes

Co-authored-by: Tom McGuire <Mcgtom10@gmail.com>